Fixed an error by changing a naïve condition
|| treats 0 like null, false or undefined, though 0 is a reasonable and valid value to expect.
Changed the condition to explicitly check for the 0 case.

Scenario - drag the thumb so that clientY is 0 (the cursor should be adjacent to the top of the viewport. An exception is thrown because touches[0] do not exist on non-touch devices. Though everything works fine afterwards, exceptions are a real downer.